refactored privacy-basics.jsx
